Prevent disbanding units on purpose to circumvent ranged unitwaittime
If unitwaittime_range >= 0, disbanding a unit is subject to
unitwaittime restrictions. Forced disbanding if a city can't afford
unit upkeep is unaffected. (If a player engineers a situation where
they have a city disband a unit on purpose just to circumvent this,
I suppose they deserve to be able to do it.)