Server option to make sure captured/bribed units always get a home city
Add server option 'homecaughtunits_always' (boolean, default disabled).  If
it is set, all captured or bribed units are given a home city (unless the
player has none).  Use the acting unit's home, or if it's homeless, use the
nearest city.

This is meant to reduce some exploits where capturing is used to make units
free of upkeep.